26-FE: Move `wrapInAPIResponse` to separate file
